+++ models/dao/transactions/ethereum/dao_transactions_ethereum_eth.sql
@@ -84,7 +84,7 @@ SELECT
     t.tx_type,
     t.token as asset_contract_address,
     'ETH' asset,
-    CAST(t.value AS DECIMAL(38,0)) as raw_value, 
+    CAST(t.value AS DOUBLE) as raw_value,
     t.value/POW(10, 18) as value, 
     t.value/POW(10, 18) * COALESCE(p.price, dp.median_price) as usd_value, 
     t.tx_hash,
