+++ models/stealcam/arbitrum/stealcam_arbitrum_events.sql
@@ -29,12 +29,12 @@ SELECT 'arbitrum' AS blockchain
 , 'Stealcam' AS collection
 , sc.id AS token_id
 , 'erc721' AS token_standard
-, CAST(1 AS DECIMAL(38,0)) AS number_of_items
+, CAST(1 AS DOUBLE) AS number_of_items
 , 0x82af49447d8a07e3bd95bd0d56f35241523fbab1 AS currency_contract
 , 'ETH' AS currency_symbol
-, CAST(sc.value AS DECIMAL(38,0)) AS amount_raw
-, CAST(sc.value/POWER(10, 18) AS DECIMAL(38,0)) AS amount_original
-, CAST(pu.price*sc.value/POWER(10, 18) AS DECIMAL(38,0)) AS amount_usd
+, CAST(sc.value AS DOUBLE) AS amount_raw
+, CAST(sc.value/POWER(10, 18) AS DOUBLE) AS amount_original
+, CAST(pu.price*sc.value/POWER(10, 18) AS DOUBLE) AS amount_usd
 , sc.contract_address AS project_contract_address
 , CAST(NULL AS string) AS aggregator_name
 , CAST(NULL AS string) AS aggregator_address
